Split logic for wikibase/auth to allow different wikibases
This ended up requiring more changes than antecipated.

Most of the work was around the functionality of the Client/Token.
Client now needs to take both token and the wikibase model it is
connecting to.

Also included: changes in test setup. uses now pytest to run all
tests. This has the added benefit of being able to use environment
variables specific for the test