Remove depends_on for docker services
During development, having services not being able to start
are more of a nuisance than anything. For example, none of the
services will be able to come up if the migrations fail at the
start, and this prevents us to see the error message *and* to
run the process to fix the migrations.

This commit also adds a 'dev.yml' compose definition that can be
used to override some settings. E.g, you can use to determine the
location of the docker volumes (useful if you don't want to have
mariadb data stored on the partition for the default location in docker
volumes)