Create 'proper' label model, use them to command label information
When presenting the information about an entity, we need to pull its
label and all related attributes (properties, qualifiers, references,
etc).

Instead of relying on a temporary attribute to inject this information
in the BatchCommand instances, we can now use a Label model and have a
many-to-many relationship.

Eventually, we can make this fetching logic smarter and avoid API
requests when/if we already have the label data, but given that this
will only happen at preview time, it should be ok.